requirements.txt
# requirements.txt
Flask>=2.0
Flask-SQLAlchemy>=2.5
python-dotenv>=0.19
Flask-Login>=0.5
email-validator>=1.1 # Often needed for form validation/Flask-Login
bcrypt>=3.2
pandas>=1.3       # For analytics
matplotlib>=3.4   # For plotting analytics